improve support for double sequences

+++ com.oracle.truffle.r.runtime/src/com/oracle/truffle/r/runtime/data/RDataFactory.java
@@ -183,7 +183,7 @@ public final class RDataFactory {
 
     public static RDoubleSequence createAscendingRange(double start, double end) {
         assert start <= end;
-        return traceDataCreated(new RDoubleSequence(start, 1, ((int) (end - start) + 1)));
+        return traceDataCreated(new RDoubleSequence(start, 1, (int) end - (int) start + 1));
     }
 
     public static RDoubleSequence createDescendingRange(double start, double end) {
